Refactor for better typing.
