+++ src/features/campaigns/types.ts
@@ -32,13 +32,8 @@ export type TaskActivity = CampaignActivityBase & {
   kind: ACTIVITIES.TASK;
 };
 
-export type MultiDayEvent = ZetkinEvent & {
-  originalEndTime: string;
-  originalStartTime: string;
-};
-
 export type EventActivity = CampaignActivityBase & {
-  data: ZetkinEvent | MultiDayEvent;
+  data: ZetkinEvent;
   kind: ACTIVITIES.EVENT;
 };
